requirements.txt
opencv-python
Pillow
pygetwindow
scikit-learn